refactor: Merge UGC services
