+++ environments/ugc_service
@@ -11,4 +11,4 @@ KAFKA_VIEW_TOPIC=views
 CLICKHOUSE_HOST=clickhouse-node1
 CLICKHOUSE_PORT=9000
 BACKOFF_MAX_TIME=300
-authjwt_secret_key=${JWT_SECRET_KEY}
+MONGODB_URI=mongodb://mongo:27017
