--- src/core/config.py
@@ -1,9 +1,0 @@
-from pydantic import BaseSettings
-
-
-class Settings(BaseSettings):
-    mongodb_uri: str
-    authjwt_secret_key: str
-
-
-settings = Settings()
