+++ ugc_service/src/api/auth.py
