+++ ugc_service/src/api/utils.py
