+++ ugc_service/src/db/mongo.py
