+++ ugc_service/src/models/__init__.py
@@ -1,0 +1,4 @@
+from models.users import User
+from models.ratings import Rating, OverallRating
+from models.reviews import Review, ReviewVote
+from models.bookmarks import Bookmark
