+++ ugc_service/src/models/base.py
@@ -1,0 +1,12 @@
+import orjson
+from pydantic import BaseModel as PydanticBaseModel
+
+
+def orjson_dumps(v, *, default):
+    return orjson.dumps(v, default=default).decode()
+
+
+class BaseModel(PydanticBaseModel):
+    class Config:
+        json_loads = orjson.loads
+        json_dumps = orjson_dumps
