+++ ugc_service/src/models/users.py
@@ -1,0 +1,7 @@
+from uuid import UUID
+
+from models.base import BaseModel
+
+
+class User(BaseModel):
+    id: UUID
