+++ ugc_service/src/services/exceptions.py
