+++ ugc_service/src/services/ratings.py
